requirements.txt
django==2.1.5
Pillow
djangorestframework
djongo
django-cors-headers
coreapi
django-filter

